Simplify directory structure
